fix(chat): guard placeholder runtime + extract E2E helpers (#199)
Guard AssistantRuntimeProvider against the placeholder sentinel runtime.
Extract shared E2E helpers (login, getSmithersAgentId, waitForOpenClawConnected)
to e2e/integration/helpers.ts to reduce duplication across specs.
