+++ packages/web/e2e/integration/15-stream-persistence.spec.ts
@@ -1,59 +1,18 @@
 // packages/web/e2e/integration/15-stream-persistence.spec.ts
 import { test, expect } from "@playwright/test";
-import type { Page, BrowserContext } from "@playwright/test";
+import type { BrowserContext } from "@playwright/test";
 import {
   FAKE_OLLAMA_SLOW_STREAM_TRIGGER,
   FAKE_OLLAMA_SLOW_STREAM_RESPONSE,
   FAKE_OLLAMA_SLOW_STREAM_DELAY_MS,
 } from "./fake-ollama-server";
+import { login, getSmithersAgentId, waitForOpenClawConnected } from "./helpers";
 
 const RESPONSE_WORDS = FAKE_OLLAMA_SLOW_STREAM_RESPONSE.split(" ");
 const FIRST_WORD = RESPONSE_WORDS[0]!;
 const LAST_WORD = RESPONSE_WORDS[RESPONSE_WORDS.length - 1]!;
 
 test.describe("Stream persistence — #199 Layer A + B end-to-end", () => {
-  async function login(page: Page) {
-    const setup = await page.request.post("/api/setup", {
-      data: {
-        name: "Integration Admin",
-        email: "admin@integration.local",
-        password: "integration-password-123",
-      },
-    });
-    expect([201, 403]).toContain(setup.status());
-
-    await page.goto("/login");
-    await page.getByLabel(/email/i).fill("admin@integration.local");
-    await page.getByLabel("Password", { exact: true }).fill("integration-password-123");
-    await page.getByRole("button", { name: /sign in/i }).click();
-    await expect(page).toHaveURL(/\/chat\//, { timeout: 15000 });
-  }
-
-  async function getSmithersAgentId(page: Page) {
-    const res = await page.request.get("/api/agents");
-    const agents = await res.json();
-    const smithers = agents.find((a: { name: string }) => a.name === "Smithers");
-    expect(smithers).toBeTruthy();
-    return smithers.id as string;
-  }
-
-  async function waitForOpenClawConnected(page: Page, timeoutMs = 120000) {
-    const deadline = Date.now() + timeoutMs;
-    let connectedSince: number | null = null;
-    while (Date.now() < deadline) {
-      const health = await page.request.get("/api/health/openclaw");
-      const data = await health.json();
-      if (data.connected) {
-        connectedSince ??= Date.now();
-        if (Date.now() - connectedSince >= 5000) return;
-      } else {
-        connectedSince = null;
-      }
-      await new Promise((r) => setTimeout(r, 500));
-    }
-    throw new Error(`OpenClaw did not connect within ${timeoutMs}ms`);
-  }
-
   test("user msg + reply survive mid-stream context-close", async ({ browser }) => {
     // First context: send a slow-streaming message, observe first token, then close.
     const ctx1: BrowserContext = await browser.newContext();
